# api/openapi/bff-web.yaml
openapi: 3.1.0
info:
  title: bff-web
  version: 1.0.0
  summary: The storefront's HTTP API.
  description: |
    One endpoint per screen, not per entity: `GET /api/v1/products` answers the
    listing grid, and a field no card renders is left behind rather than
    forwarded because a service happened to return it. A shape here is a screen's
    shape, and it changes when the screen does.

    This document is hand-written and checked against the running router by a
    contract test, so an endpoint or a field that appears in one and not the
    other fails the build rather than drifting quietly.

    ## Reading a failure

    Every failure this API decides on is one shape — see `ErrorResponse` — and
    every field is camelCase.

    Branch on `error.code`, never on the HTTP status. A 409 alone cannot say
    whether an email was taken or an order was already paid, and the codes are
    machine-stable: renaming one is a breaking change.

    Three answers come from the gateway instead and cannot carry that shape,
    because it short-circuits them before this service is reached. They are JSON,
    but they carry `message` and no `error.code`:

    | Status | When |
    | --- | --- |
    | 429 | rate limited at the edge |
    | 413 | body past the gateway's hard cap, 8× the one this service enforces |
    | 502, 503, 504 | this service is unreachable or too slow to answer |

    So a client needs one rule for the remainder: **no `error.code` means derive
...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/v1/products:
    get:
      operationId: listProducts
      tags: [storefront]
      summary: One page of the catalog, as a grid of cards
      description: |
        Anonymous: browsing is what a shopper does before deciding to have an
        account.

        Only products on sale are returned, and there is no parameter that can
        ask for anything else — drafts and archived products are an admin
        console's business, not this API's. A query parameter this endpoint does
        not name is ignored rather than rejected, so a campaign link carrying its
        own tracking parameters is still a valid request.
      parameters:
        - name: category
          in: query
          required: false
          description: Exact match on a category slug. Every category when absent.
          schema:
            type: string
            maxLength: 64
          example: clothing/shirts
        - name: pageSize
          in: query
          required: false
          description: |
            How many cards to return. Omit it, or send 0, to let the catalog
            choose.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 100
        - name: pageToken
          in: query
          required: false
          description: |
            The `nextPageToken` from the previous response. Opaque: read nothing
            from it and construct nothing into it.
          schema:
            type: string
            maxLength: 512
      responses:
        '200':
          description: One page. An empty page is an empty array, never null.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/UpstreamUnavailable'
        '504':
          $ref: '#/components/responses/UpstreamTimeout'
  /api/v1/products/{id}:
    get:
      operationId: getProduct
      tags: [storefront]
      summary: One product, as its own page
      description: |
        Anonymous, for the reason the listing is.

        Only published products are answered for. A draft, an archived product,
        and an id nobody ever minted are all `404 PRODUCT_NOT_FOUND`, and the
        three are deliberately indistinguishable — otherwise the status code
        alone would tell anyone walking UUIDs which ones name an unreleased
        product.

        Availability is the one optional part of this page. When the warehouse
        does not answer, every variant's `availableQuantity` is `null` and
        nothing else about the response changes; a client that renders `null` as
        "sold out" is wrong in the direction that costs a sale.
      parameters:
        - name: id
          in: path
          required: true
          description: The product's id, as `products.id` from the listing.
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
          example: 3f2504e0-4f89-41d3-9a0c-0305e82c3301
      responses:
        '200':
          description: The product, with every way it can be bought.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: |
            No product on sale under this id — `PRODUCT_NOT_FOUND`.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationID'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - properties:
                      error:
                        properties:
                          code:
                            enum: [PRODUCT_NOT_FOUND]
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/UpstreamUnavailable'
        '504':
          $ref: '#/components/responses/UpstreamTimeout'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The access token from `POST /api/v1/auth/login`, as
        `Authorization: Bearer <token>`. It is an ES256 JWT and lives 15 minutes;
        renew it with `POST /api/v1/auth/refresh` rather than signing in again.

        `X-User-ID` and `X-User-Roles` are ignored wherever they appear. Identity
        comes from the signature this service verified itself.
  headers:
    CorrelationID:
      description: |
        The ID tying this response to every log line and span behind it. Present
        on every answer, this API's and the gateway's alike.
      schema:
        type: string
  responses:
    BadRequest:
      description: |
        The request could not be read or broke a rule.

        A body that will not decode is a bug in the client and carries a reason
        code with no translated prose: `EMPTY_BODY`, `MALFORMED_JSON`,
        `UNKNOWN_FIELD`, `INVALID_FIELD_TYPE`. A request that decodes but breaks
        a rule is a person's mistake and comes back as `INVALID_INPUT` with
        `error.fields`, one entry per failing field, each already translated.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
        Content-Language:
          description: The language `error.fields[].message` was written in.
          schema:
            type: string
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  error:
                    properties:
                      code:
                        enum:
                          - INVALID_INPUT
                          - MALFORMED_JSON
                          - EMPTY_BODY
                          - UNKNOWN_FIELD
                          - INVALID_FIELD_TYPE
...
    InternalError:
      description: |
        Something here broke — `INTERNAL`, or `UNKNOWN` and `DATA_LOSS` for a
        failure that could not even say that much. All three mean the same thing
        to a client: nothing it sent was wrong and nothing it can change will
        help. The message is fixed prose and the detail is in the logs under
        this response's correlation ID.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  error:
                    properties:
                      code:
                        enum: [INTERNAL, UNKNOWN, DATA_LOSS]
    UpstreamUnavailable:
      description: |
        A service this endpoint depends on could not be reached — `UNAVAILABLE`.

        A 503 with no `error.code` at all came from the gateway instead, and
        means this service is the one that could not be reached. Both are
        `UPSTREAM_UNAVAILABLE` as far as a client's handling goes.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  error:
                    properties:
                      code:
                        enum: [UNAVAILABLE]
    UpstreamTimeout:
      description: |
        A service this endpoint depends on did not answer inside the request's
        budget — `DEADLINE_EXCEEDED`. Retrying is reasonable; retrying
        immediately is not.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationID'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  error:
                    properties:
                      code:
                        enum: [DEADLINE_EXCEEDED]
  schemas:
...
    ProductsResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        nextPageToken:
          type: string
          description: |
            Absent on the last page. Send it back as `pageToken` to ask for the
            next one.
    ProductCard:
      type: object
      description: |
        One product as a grid of cards needs it, which is much less than the
        catalog holds about it. The description and the variant list are
        deliberately absent: twenty cards would carry twenty descriptions and
        every SKU to render text no card shows.
      required: [id, name, category, priceFrom, variantCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        priceFrom:
          description: |
            The cheapest variant's price, which is what "from ฿x" on a card
            quotes. Null for a product with no variants — a state the storefront
            should never see, since publishing requires one.
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
        variantCount:
          type: integer
          description: |
            How many ways this product can be bought, so a card knows whether
            `priceFrom` is the price or the lowest of several.
    ProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: '#/components/schemas/ProductDetail'
    ProductDetail:
      type: object
      description: |
        One product as its own page needs it, which is everything a card leaves
        out: the words, and every way the thing can be bought.

        There is no status field. Only published products are ever returned
        here, so it would carry the same value on every response.
      required: [id, name, description, category, variants]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        variants:
          type: array
          description: |
            Never empty for a published product — publishing requires one.
          items:
            $ref: '#/components/schemas/ProductVariant'
    ProductVariant:
      type: object
      description: |
        One buyable unit: what it costs, what distinguishes it from its
        siblings, and whether there is one left.
      required: [id, sku, price, attributes, availableQuantity]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          description: |
            What the warehouse, a picking list, and a supplier's invoice all
            call this thing. Uppercase, and stable.
          example: SHIRT-OX-M
        price:
          oneOf:
            - $ref: '#/components/schemas/Money'
            - type: 'null'
          description: |
            Null for a variant the catalog holds no price for — a state a
            published product should not be in, and one a zero would describe
            wrongly.
        attributes:
          type: object
          description: |
            Whatever the catalog was told distinguishes this variant. Opaque:
            read a key only if you put it there. `{}` when there are none.
          additionalProperties:
            type: string
          example:
            size: M
        availableQuantity:
          description: |
            How many may still be bought. `0` means the warehouse has none;
            `null` means it did not answer and nobody knows — render the two
            differently, because treating `null` as sold out hides a product
            that is on the shelf.

            It is the count and not a flag: whether to say "only 3 left" or just
            "in stock" is the client's to decide, and any rule about what a
            shopper may learn belongs to the service that owns the number.
          oneOf:
            - type: integer
              format: int32
              minimum: 0
            - type: 'null'
    Money:
      type: object
      description: |
        An amount in a currency's minor unit — 1050 is THB 10.50 — with the
        currency saying how many decimal places that is. Formatting is the
        client's: the locale an amount is rendered in is the browser's, not this
        API's.
      required: [amountMinor, currencyCode]
      properties:
        amountMinor:
          type: integer
          format: int64
        currencyCode:
          type: string
          description: ISO 4217.
          minLength: 3
          maxLength: 3
      example:
        amountMinor: 89000
        currencyCode: THB
    ErrorResponse:
      type: object
      description: The body behind every failure this API decided on.
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            The machine-stable reason to branch on, finer-grained than the HTTP
            status on purpose. Renaming one is a breaking change.

            A failure raised below the handlers — a refused connection, a
            tripped breaker, a spent deadline — has no reason of its own, and
            the name of the status it arrived as stands in: `UNAVAILABLE`,
            `DEADLINE_EXCEEDED`. Every response below lists the codes it can
            carry.
          examples:
            - EMAIL_ALREADY_REGISTERED
            - INVALID_CREDENTIALS
            - TOKEN_EXPIRED
            - INVALID_INPUT
        message:
          type: string
          description: |
            English prose, for a developer reading a log or a network tab. Never
            render it to a user: it is untranslated, and a 5xx replaces it with
            fixed text anyway so that nothing internal leaks.
        fields:
          type: array
          description: Set when the request broke per-field validation.
          items:
            $ref: '#/components/schemas/FieldError'
        metadata:
          type: object
          description: |
            The facts behind `code` — the SKU that was short, the status an order
            was actually in — so a client can compose its own message without
            parsing one.
          additionalProperties:
            type: string
        correlationId:
          type: string
          description: |
            Present when this service wrote the body. Prefer the
            `X-Correlation-ID` header, which is on every answer including the
            gateway's.
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: |
            The path as the client wrote it, in the same camelCase the request
            used, with indexes for collections — `items[0].quantity`.
          examples: [pageSize, email]
        code:
          type: string
          description: The rule that failed, for a client that wants its own wording.
          examples: [required, email, lte]
        message:
          type: string
          description: |
            The failure in the language `Accept-Language` asked for, ready to
            render next to the input. The only translated string in this API.